Move shared components to shared-components folder, and update imports
