+++ src/app/app.component.ts
@@ -3,7 +3,7 @@ import { Subject } from 'rxjs';
 import { Router } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 import { RouterOutlet } from '@angular/router';
-import { IdleTimeoutComponent } from './game/idle-timeout/idle-timeout.component';
+import { IdleTimeoutComponent } from './game/shared-components/idle-timeout/idle-timeout.component';
 import { routeTransitionAnimations } from './route-transition-animations';
 import { environment } from '../environments/environment';
 import { GameStateService } from './game/services/game-state-service';
