+++ src/app/game/shared-components/speech-bubble/speech-bubble.component.html
