+++ src/app/game/shared-components/speech-bubble/speech-bubble.component.scss
@@ -1,4 +1,4 @@
-@import '../../../variables.scss';
+@import '../../../../variables.scss';
 
 * {
   text-align: center;
